+++ particlegenerator.h
@@ -9,7 +9,7 @@ class ParticleGenerator
 public:
     ParticleGenerator();
     virtual ~ParticleGenerator();
-    virtual void generate(float dt, ParticleDataContainer *p, MultithreadOverseer *task_menager, size_t start_id, size_t end_id) = 0;
+    virtual void generate(double dt, ParticleDataContainer *p, MultithreadOverseer *task_menager, size_t start_id, size_t end_id) = 0;
 };
 
 #endif // PARTICLEGENERATOR_H
